requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
numpy>=1.20.0
Pillow>=9.0.0
pydicom>=2.3.0
scikit-image>=0.18.0
jupyter>=1.0.0
notebook>=6.4.0
